cleaned bot layer
